Completed the admin panel.

+++ dao/src/main/java/hr/algebra/dao/rss/AssetService.java
@@ -10,6 +10,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.UUID;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 public class AssetService {
     private final Path FOLDER = Paths.get("assets");
@@ -41,6 +43,10 @@ public class AssetService {
     }
 
     public void removeImage(String imagePath) throws AssetException {
+        if(imagePath == null) {
+            return;
+        }
+
         Path file = Paths.get(imagePath);
         try {
             Files.delete(file);
